requirements.txt
requests==2.24.0
selenium==3.141.0
Pillow==8.2.0
pytesseract==0.3.7